feature: add localization
This is a very long commit, while implementing the card localization
feature I faced some problems. The first one was out of memory during
assets download, which I limited to 2 gorotines only.

The second one is that I tried to use in-memory SQLite as database, but
the RAM consumption spiked on every request while querying the database.

I ended up with mongo atlas free tier to save card assets and it's
working now.